+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 107.0.7 (Apr 8, 2023)
+
+### Patch
+
+- Docs: 4/7 weekly digest ([#2814](https://github.com/pinterest/gestalt/pull/2814)) - [Preview link](https://deploy-preview-2814--gestalt.netlify.app)
+
+## 107.0.6 (Apr 7, 2023)
+
+### Patch
+
+- Updated mobile and documentation statuses for all components ([#2804](https://github.com/pinterest/gestalt/pull/2804)) - [Preview link](https://deploy-preview-2804--gestalt.netlify.app)
+
+## 107.0.5 (Apr 7, 2023)
+
+### Patch
+
+- Docs: fix mobile examples display ([#2811](https://github.com/pinterest/gestalt/pull/2811)) - [Preview link](https://deploy-preview-2811--gestalt.netlify.app)
+
 ## 107.0.4 (Apr 6, 2023)
 
 ### Patch